Add 3rd option for user feedback match success
other than 'no match' or 'good match', users can now say that it was a multi-phase project and the wrong match they got was ust because of that. This is option `2` (other than `0` or `1`). This adds column `multi_phase` to df_matched and logs a 1 under that column if option `2` is selected and results in the match being ignored when building the train set. Otherwise, option `2` is simply handled as option `0` for all other matters.